refactor: expose initTheme from utils
