+++ apps/shop-abc/src/app/layout.tsx
@@ -1,7 +1,7 @@
 // src/app/layout.tsx
 import "./globals.css";
 import { CartProvider } from "@/contexts/CartContext";
-import { initTheme } from "@platform-core/src/utils/initTheme";
+import { initTheme } from "@platform-core/utils";
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
