+++ packages/platform-core/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { getShopFromPath } from "./getShopFromPath";
 export { replaceShopInPath } from "./replaceShopInPath";
 export { fillLocales } from "./locales";
+export { initTheme } from "./initTheme";
